r3v: print the subresource rowPitch through its VkDeviceSize width
VkImageSubresourceLayout.rowPitch is a VkDeviceSize, so the NDEBUG
fall-through message in r3v_native_transfer_ops_test.c formatted a
64-bit value through %u; clang -Wformat under -Werror refuses the
translation unit, which leaves every r3v test unbuilt.  The message now
widens the value to unsigned long long and prints it through %llu, the
same shape the neighboring offset and size arguments use.

+++ src/amd/r300/vulkan/tests/r3v_native_transfer_ops_test.c
@@ -810,9 +810,10 @@ check_optimal_subresource_layout_oracle(const struct fixture *f)
                   layout.size == expect_size,
                "under NDEBUG, vkGetImageSubresourceLayout on the "
                "OPTIMAL image falls through to the transfer family's "
-               "linear span (offset %llu rowPitch %u size %llu, "
+               "linear span (offset %llu rowPitch %llu size %llu, "
                "expected offset 0 rowPitch %u size %llu)",
-               (unsigned long long)layout.offset, layout.rowPitch,
+               (unsigned long long)layout.offset,
+               (unsigned long long)layout.rowPitch,
                (unsigned long long)layout.size, expect_row_pitch,
                (unsigned long long)expect_size);
       }
